+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 
 * New `vcov.bhm()` function returns the covariance matrix estimated at the posterior's mode.
+* New `logposterior()` function returns the log-posterior function for the model's data.
 
 
 # homeR 0.2